Resolves 33 textual conflicts and one modify/delete (mod_cdr_mongodb) by
favouring Telnyx changes, with documented exceptions for stale module
references upstream removed from tree, the v1.11.0 version bump itself,
and additive upstream content that does not contradict Telnyx code.

Also fixes two pre-existing TEL-6154 leftovers surfaced by the build:
mod_esf and mod_skinny call sites still passed the removed bundle_*_port
args to switch_rtp_new. Removes one duplicate case label pair in
mod_sofia (TEL-6404 + upstream-added cases for
SWITCH_MESSAGE_INDICATE_TRANSCODING_NECESSARY / RESAMPLE_EVENT).

+++ src/mod/applications/mod_http_cache/mod_http_cache.c
@@ -1310,10 +1310,10 @@ static switch_status_t http_get(url_cache_t *cache, http_profile_t *profile, cac
 #endif
 		int i;
 
-		switch_curl_easy_setopt(curl_handle, CURLOPT_FOLLOWLOCATION, 1);
-		switch_curl_easy_setopt(curl_handle, CURLOPT_MAXREDIRS, 10);
-		switch_curl_easy_setopt(curl_handle, CURLOPT_FAILONERROR, 1);
-		switch_curl_easy_setopt(curl_handle, CURLOPT_NOSIGNAL, 1);
+		switch_curl_easy_setopt(curl_handle, CURLOPT_FOLLOWLOCATION, (long)1);
+		switch_curl_easy_setopt(curl_handle, CURLOPT_MAXREDIRS, (long)10);
+		switch_curl_easy_setopt(curl_handle, CURLOPT_FAILONERROR, (long)1);
+		switch_curl_easy_setopt(curl_handle, CURLOPT_NOSIGNAL, (long)1);
 		if (headers) {
 			switch_curl_easy_setopt(curl_handle, CURLOPT_HTTPHEADER, headers);
 		}
